requirements.txt
uvicorn[standard]>=0.37.0
fastapi==0.118.0

numpy==2.3.3
opencv_python==4.11.0.86
opencv_python_headless==4.10.0.84
python-dotenv==1.1.1
Requests==2.32.5
ultralytics==8.3.205
python-multipart==0.0.20
